// src/enums/cruds/sellingFields.enum.ts
export enum SellingField {
  SaleType = 'saleType',
  SaleDate = 'saleDate',
  AliveWeight = 'aliveWeight',
  DeadWeight = 'deadWeight',
  Quantity = 'quantity',
  UnityValue = 'unityValue',
  ArrobaCost = 'arrobaCost',
  CarcassYield = 'carcassYield',
  Category = 'category',
  Retreat = 'retreat',
  Area = 'area',
  Observation = 'observation',
  DestinationFarm = 'destinationFarm',
  DestinationRetreat = 'destinationRetreat',
  DestinationArea = 'destinationArea',
  IsExternalDestination = 'isExternalDestination',
  Age = 'age',
  AnimalLotId = 'animalLotId',
  FatteningSystemId = 'fatteningSystemId',
  HarvestConfiguration = 'harvestConfiguration',
}